Remove relation database.
