+++ transactron/core/__init__.py
@@ -1,5 +1,4 @@
 from .typing import *  # noqa: F401
-from .relation_database import *  # noqa: F401
 from .transaction_base import *  # noqa: F401
 from .modules import *  # noqa: F401
 from .sugar import *  # noqa: F401
