+++ transactron/core/schedulers.py
@@ -1,16 +1,13 @@
 from amaranth import *
+from typing import TYPE_CHECKING
 from .._utils import Scheduler
-from .relation_database import MethodMap
 from .typing import TransactionGraph, TransactionGraphCC, PriorityOrder
 
-__all__ = [
-    "eager_deterministic_cc_scheduler",
-    "trivial_roundrobin_cc_scheduler",
-]
-
+if TYPE_CHECKING:
+    from .manager import MethodMap
 
 def eager_deterministic_cc_scheduler(
-    method_map: MethodMap, gr: TransactionGraph, cc: TransactionGraphCC, porder: PriorityOrder
+    method_map: "MethodMap", gr: TransactionGraph, cc: TransactionGraphCC, porder: PriorityOrder
 ) -> Module:
     """eager_deterministic_cc_scheduler
 
@@ -48,7 +45,7 @@ def eager_deterministic_cc_scheduler(
 
 
 def trivial_roundrobin_cc_scheduler(
-    method_map: MethodMap, gr: TransactionGraph, cc: TransactionGraphCC, porder: PriorityOrder
+    method_map: "MethodMap", gr: TransactionGraph, cc: TransactionGraphCC, porder: PriorityOrder
 ) -> Module:
     """trivial_roundrobin_cc_scheduler
